refactor(processing): rename MarkEvidenceSubmissionService to MarkEvidenceSubmissionSubmitService and update references; adjust ReviewAutoConfiguration package path
