+++ review/src/main/java/pt/estga/review/config/ReviewAutoConfiguration.java
@@ -1,4 +1,4 @@
-package pt.estga.review;
+package pt.estga.review.config;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.persistence.autoconfigure.EntityScan;
