+++ review/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-pt.estga.review.ReviewAutoConfiguration
+pt.estga.review.config.ReviewAutoConfiguration
